Move FF related constants to appropriate class
As a side note, I'm smelling the Constant modules a little, we can
probably clean this up later.  It would probably be better to have
constants available in the modules that aren't prefixed with a vendor
name, because it will cut out subclass methods that are just one
liners going for the constant.