Extract logic to write attribute to prevent double checking for the attribute being valid
When initializing the config from hash we need to check if it's part of
the known list of attributes, and the same happens when writing the
attribute (I'm still wondering why we need to do that check, since the
methods are initialized from the list of attributes, but decided not to
touch it for now).

Extracting this logic allows us to bypass the duplicated inclusion check
when initializing from hash. It also bypasses `send` which helps speed
things up.