Remove caching and make overrides work as expected
We ran into a bug where overrides weren't working as expected. The user
expectation is that an override should override a config value for whatever
the current configuration for the current request. But, as implemented, they
actually were just forks of a base config..regardless of whether that base
config changed during runtime. So, overrides no work that way and we don't
currently support any way of using an alternate config. This might be added
back later if it is useful.

While making the changes related to overrides it became obvious that caching
was a huge source of complexity. So, this also removes all caching. A quick
local benchmark shows that a non-cached config takes about .18ms to build out
a set of headers, so it shouldn't be a big concern.