Add test to exercise override opting out without default_src
On secure_headers 5.x it was possible to override the CSP directives
when optin out without having to define a default_src.

Now on 6.x it is required to set the default_src when overriding other
directives.

It is not clear in the CHANGELOG/upgrade guide if this change is by
design or if it is just a side effect of other changes.

I could not find anyting in the spec that says that default_src is
required or not, so I decided to open a PR with a test to get feedback
on that.

If this is undesirable behavior I'm willing to change this PR to fix the
problem.

Let me know what are the next steps.

Thanks.